# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Olexfilms API
  version: 0.1.0
  summary: Olexfilms HTTP API — auth, RBAC, organizations, platform infrastructure
  description: |
    Olexfilms Ecosystem HTTP API (infrastructure layer of the F0 base).

    **In scope (F0 base):** authentication, global RBAC, platform users/roles,
    organizations, notifications, Web Push (VAPID), passkeys, step-up, storage,
    import/export/bulk engines, search, logs, and Centrifugo realtime tokens.

    Business modules arrive with their own Linear issues. Live Go routes
    (`backend/internal/modules/**/routes.go`) are the source of truth if a path
    below is missing from the router.

    **Roles (TEC-85, scoped RBAC):**
    - `super_admin` — platform role; assigned via `make create-super-admin`; the
      only role that may impersonate.
    - Organization roles granted per membership: `center_staff`,
      `center_warehouse`, `center_accounting`, `center_social`,
      `distributor_owner|staff|warehouse_staff|accounting`,
      `dealer_owner|staff|accounting`.
    - Global portal roles: `customer`, `fleet` (read-only).
    - Every grant has a scope: `all > brand > subtree > managed > assigned > own`
      (`customer` only covers itself). `make roles-sync` reconciles the DB with
      the Go catalog (`backend/internal/platform/rbac/catalog.go`).
    - Custom roles via `/v1/platform/roles`

    **Auth:** Bearer JWT access tokens + opaque refresh tokens.
...
    Error catalog: `backend/docs/api/errors.md`.

...
servers:
...
paths:
...
  /v1/platform/permissions:
    get:
      tags: [Platform]
      operationId: listPlatformPermissions
      summary: Permission catalog with modules and allowed scopes
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer }
        - name: offset
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Permissions
          content:
            application/json:
              schema:
                type: object
                required: [success, data, meta]
                properties:
                  success: { type: boolean, enum: [true] }
                  data:
                    type: object
                    required: [items, total, limit, offset]
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/PermissionSummary" }
                      total: { type: integer, format: int64 }
                      limit: { type: integer }
                      offset: { type: integer }
                  meta: { $ref: "#/components/schemas/ResponseMeta" }
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/platform/roles/{uuid}:
    get:
      tags: [Platform]
      operationId: getPlatformRole
      summary: Role with its scoped grants
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Role detail
          content:
            application/json:
              schema:
                type: object
                required: [success, data, meta]
                properties:
                  success: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/RoleDetail" }
                  meta: { $ref: "#/components/schemas/ResponseMeta" }
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/platform/users/{uuid}:
    get:
      tags: [Platform]
      operationId: getPlatformUser
      summary: Get platform user
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceUUID"
      responses:
        "200":
          description: User with memberships
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopePlatformUserDetail"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags: [Platform]
      operationId: patchPlatformUser
      summary: Update platform user
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceUUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchPlatformUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopePublicUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ResourceUUID:
      name: uuid
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
...
  responses:
...
    BadRequest:
      description: Validation or malformed request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Unauthenticated:
      description: Missing or invalid credentials / access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Authenticated but not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Conflict (duplicate / constraint)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    ErrorDetail:
      type: object
      required: [message]
      properties:
        field: { type: string }
        message: { type: string }
        code: { type: string }
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          examples:
            - VALIDATION_ERROR
            - INVALID_RESET_CODE
            - INVALID_VERIFICATION_CODE
            - INVALID_CREDENTIALS
            - UNAUTHENTICATED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - PASSWORD_RESET_FAILED
            - INTERNAL_ERROR
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
    ResponseMeta:
      type: object
      required: [request_id]
      properties:
        request_id: { type: string }
    Error:
      type: object
      required: [success, error, meta]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          $ref: "#/components/schemas/ErrorBody"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    PublicUser:
      type: object
      required: [uuid, email, name, surname, status, is_super_admin, email_verified]
      properties:
        uuid:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        surname:
          type: string
        status:
          type: string
          examples: [active, disabled]
        is_super_admin:
          type: boolean
        email_verified:
          type: boolean
        locale:
          description: The user's own language; null inherits from the organization (see Me.effective_locale).
          oneOf:
            - $ref: "#/components/schemas/Locale"
            - type: "null"
        timezone:
          type: [string, "null"]
          description: The user's own IANA timezone; null inherits from the organization.
          examples: ["Asia/Dubai"]
    Locale:
      type: string
      description: >-
        Canonical locale code (K10). Inputs are normalized: "zh_CN", "zh-cn"
        and "zh" become zh-CN; a regional code falls back to its language
        ("tr-TR" -> tr). Unknown codes are rejected with 422.
      enum: [tr, en, bg, de, el, uk, ru, fr, es, it, zh-CN, az, ar]
...
    PatchPlatformUserRequest:
      type: object
      properties:
        name: { type: string }
        surname: { type: string }
        status: { type: string }
        role_uuids:
          type: array
          items: { type: string, format: uuid }
...
    RoleSummary:
      type: object
      required: [uuid, name, slug]
      properties:
        uuid: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        description: { type: [string, "null"] }
        is_system: { type: boolean }
        org_type:
          type: string
          description: Subject kind of the role; organization roles are granted per membership.
          enum: [platform, center, distributor, dealer, customer, fleet]
    PermissionScope:
      type: string
      description: >-
        Reach of a grant. all > brand > subtree > managed > assigned > own;
        customer only covers itself.
      enum: [all, brand, subtree, managed, assigned, own, customer]
    PermissionSummary:
      type: object
      required: [uuid, name, slug, module, scopes, is_sensitive, super_admin_only]
      properties:
        uuid: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        module: { type: string }
        scopes:
          type: array
          items: { $ref: "#/components/schemas/PermissionScope" }
        is_sensitive: { type: boolean }
        super_admin_only: { type: boolean }
        description: { type: [string, "null"] }
    RoleGrant:
      type: object
      required: [permission, scope]
      properties:
        permission: { type: string }
        scope: { $ref: "#/components/schemas/PermissionScope" }
    RoleDetail:
      allOf:
        - $ref: "#/components/schemas/RoleSummary"
        - type: object
          required: [permission_slugs, grants]
          properties:
            permission_slugs:
              type: array
              items: { type: string }
            grants:
              type: array
              items: { $ref: "#/components/schemas/RoleGrant" }
    PlatformUserDetail:
      allOf:
        - $ref: "#/components/schemas/PublicUser"
        - type: object
          required: [roles, auth_methods]
          properties:
            roles:
              type: array
              items:
                $ref: "#/components/schemas/RoleSummary"
            auth_methods:
              type: array
              items:
                $ref: "#/components/schemas/UserAuthMethod"
    UserAuthMethod:
      type: object
      required: [kind, linked_at]
      properties:
        kind:
          type: string
          enum: [password, passkey, oauth]
        provider:
          type: string
          description: OAuth provider slug when kind is oauth
          examples: [github, google, facebook, apple]
        label:
          type: [string, "null"]
          description: Optional display label (e.g. passkey name)
        linked_at:
          type: string
          format: date-time
...
    EnvelopePublicUser:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, enum: [true] }
        data: { $ref: "#/components/schemas/PublicUser" }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
...
    EnvelopePlatformUserDetail:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, enum: [true] }
        data: { $ref: "#/components/schemas/PlatformUserDetail" }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
...